+++ samples/sampler/stories/default/PageViews.js
@@ -9,6 +9,7 @@ PageViews.displayName = 'PageViews';
 
 const Config = mergeComponentMetadata('PageViews', PageViews);
 const propOptions = {
+	pageIndicatorPosition: ['top', 'bottom'],
 	pageIndicatorType: ['dot', 'number']
 };
 
@@ -18,7 +19,7 @@ export default {
 };
 
 export const _PageViews = (args) => (
-	<PageViews arranger={BasicArranger} fullContents={args['fullContents']} pageIndicatorType={args['pageIndicatorType']}>
+	<PageViews arranger={BasicArranger} fullContents={args['fullContents']} pageIndicatorPosition={args['pageIndicatorPosition']} pageIndicatorType={args['pageIndicatorType']}>
 		<PageViews.Page aria-label="This is a description for page 1">
 			<div style={{padding: '24px', width: '50%'}}>
 				<Item>Item 1</Item>
@@ -67,6 +68,7 @@ export const _PageViews = (args) => (
 );
 
 boolean('fullContents', _PageViews, Config, false);
+select('pageIndicatorPosition', _PageViews, propOptions.pageIndicatorPosition, Config, 'bottom');
 select('pageIndicatorType', _PageViews, propOptions.pageIndicatorType, Config, 'dot');
 
 _PageViews.storyName = 'PageViews';