+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Added
 
 - `sandstone/Alert` public class names `alert`, `content`, `fullscreen`, and `title`
+- `sandstone/Pageviews` `pageIndicatorPosition` prop to provide a way to determine where to place the page indicator.
 
 ## [2.9.0] - 2024-07-17